+++ kitty/launch.py
@@ -460,6 +460,9 @@ def load_watch_modules(watchers: Iterable[str]) -> Optional[Watchers]:
         w = m.get('on_cmd_startstop')
         if callable(w):
             ans.on_cmd_startstop.append(w)
+        w = m.get('on_color_scheme_preference_change')
+        if callable(w):
+            ans.on_color_scheme_preference_change.append(w)
     return ans